remark: prim type to its wrap

+++ neobisss/src/Week4/src/main/java/org/example/onlineshop/dto/ProductDTO.java
@@ -1,11 +1,14 @@
 package org.example.onlineshop.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 @Data
 public class ProductDTO {
 
-    private long id;
+    private Long id;
     private String productName;
-    private double price;
+    private BigDecimal price;
 
 }
